Adjusted core per review remarks
