+++ src/test/java/com/microsoft/sqlserver/jdbc/bulkCopy/BulkCopyAllTypesTest.java
@@ -8,6 +8,8 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.sql.Timestamp;
+import java.sql.Types;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
@@ -16,14 +18,32 @@ import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
 
 import com.microsoft.sqlserver.jdbc.ComparisonUtil;
+import com.microsoft.sqlserver.jdbc.RandomData;
+import com.microsoft.sqlserver.jdbc.RandomUtil;
 import com.microsoft.sqlserver.jdbc.SQLServerBulkCopy;
+import com.microsoft.sqlserver.jdbc.SQLServerBulkCopyOptions;
 import com.microsoft.sqlserver.jdbc.TestUtils;
+import com.microsoft.sqlserver.testframework.AbstractSQLGenerator;
 import com.microsoft.sqlserver.testframework.AbstractTest;
 import com.microsoft.sqlserver.testframework.Constants;
 import com.microsoft.sqlserver.testframework.DBConnection;
 import com.microsoft.sqlserver.testframework.DBStatement;
 import com.microsoft.sqlserver.testframework.DBTable;
 import com.microsoft.sqlserver.testframework.PrepUtil;
+
+import javax.sql.RowSetMetaData;
+import javax.sql.rowset.CachedRowSet;
+import javax.sql.rowset.RowSetFactory;
+import javax.sql.rowset.RowSetMetaDataImpl;
+import javax.sql.rowset.RowSetProvider;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 
 @RunWith(JUnitPlatform.class)
@@ -35,6 +55,7 @@ public class BulkCopyAllTypesTest extends AbstractTest {
     @BeforeAll
     public static void setupTests() throws Exception {
         setConnection();
+        createDatetimeTestTable();
     }
 
     /**
@@ -98,6 +119,75 @@ public class BulkCopyAllTypesTest extends AbstractTest {
         try (Statement stmt = connection.createStatement()) {
             TestUtils.dropTableIfExists(tableSrc.getEscapedTableName(), stmt);
             TestUtils.dropTableIfExists(tableDest.getEscapedTableName(), stmt);
+            TestUtils.dropTableIfExists(dateTimeTestTable, stmt);
+        }
+    }
+
+    public static final int DATETIME_COL_COUNT = 2;
+    public static final int DATETIME_ROW_COUNT = 1;
+    private static final String dateTimeTestTable =
+            AbstractSQLGenerator.escapeIdentifier(RandomUtil.getIdentifier("bulkCopyTimestampTest"));
+
+    @Test
+    public void testBulkCopyTimestamp() throws SQLException {
+        List<Timestamp> timeStamps = new ArrayList<>();
+        try (Connection con = getConnection(); Statement stmt = connection.createStatement()) {
+            RowSetFactory rsf = RowSetProvider.newFactory();
+            CachedRowSet crs = rsf.createCachedRowSet();
+            RowSetMetaData rsmd = new RowSetMetaDataImpl();
+            rsmd.setColumnCount(DATETIME_COL_COUNT);
+
+            for (int i = 1; i <= DATETIME_COL_COUNT; i++) {
+                rsmd.setColumnName(i, String.format("c%d", i));
+                rsmd.setColumnType(i, Types.TIMESTAMP);
+            }
+            crs.setMetaData(rsmd);
+
+            for (int i = 0; i < DATETIME_COL_COUNT; i++) {
+                timeStamps.add(RandomData.generateDatetime(false));
+            }
+
+            for (int ri = 0; ri < DATETIME_ROW_COUNT; ri++) {
+                crs.moveToInsertRow();
+
+                for (int i = 1; i <= DATETIME_COL_COUNT; i++) {
+                        crs.updateTimestamp(i, timeStamps.get(i - 1));
+                }
+                crs.insertRow();
+            }
+            crs.moveToCurrentRow();
+
+            try (SQLServerBulkCopy bcOperation = new SQLServerBulkCopy(con)) {
+                SQLServerBulkCopyOptions bcOptions = new SQLServerBulkCopyOptions();
+                bcOptions.setBatchSize(5000);
+                bcOperation.setDestinationTableName(dateTimeTestTable);
+                bcOperation.setBulkCopyOptions(bcOptions);
+                bcOperation.writeToServer(crs);
+            }
+
+            try (ResultSet rs = stmt.executeQuery("select * from " + dateTimeTestTable)) {
+                assertTrue(rs.next());
+
+                for (int i = 1; i <= DATETIME_COL_COUNT; i++) {
+                    long expectedTimestamp = getTime(timeStamps.get(i - 1));
+                    long actualTimestamp = getTime(rs.getTimestamp(i));
+
+                    assertEquals(expectedTimestamp, actualTimestamp);
+                }
+            }
+        }
+    }
+
+    private static long getTime(Timestamp time) {
+        return (3 * time.getTime() + 5) / 10;
+    }
+
+    private static void createDatetimeTestTable() throws SQLException {
+        try (Statement stmt = connection.createStatement()) {
+            String colSpec = IntStream.range(1, DATETIME_COL_COUNT + 1).mapToObj(x -> String.format("c%d datetime", x)).collect(
+                    Collectors.joining(","));
+            String sql1 = String.format("create table %s (%s)", dateTimeTestTable, colSpec);
+            stmt.execute(sql1);
         }
     }
 }
